--- cc/api_user_list.go
@@ -1,179 +1,0 @@
-/*
-IdentityNow cc (private) APIs
-
-No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
-
-API version: 1.0.0
-*/
-
-package cc
-
-import (
-	"bytes"
-	"context"
-	"encoding/json"
-	"io/ioutil"
-	"net/http"
-	"net/url"
-)
-
-// ListUserListApiService SourcesAggregationApi service
-type ListUserListApiService service
-
-type ApiListUserListRequest struct {
-	ctx        context.Context
-	ApiService *ListUserListApiService
-	limit      *int32
-	offset     *int32
-	count      *bool
-	sorters    *[]QuerySort
-	filters    *string
-}
-
-// Max number of results to return. See [V3 API Standard Collection Parameters](https://developer.sailpoint.com/idn/api/standard-collection-parameters) for more information.
-func (r ApiListUserListRequest) Limit(limit int32) ApiListUserListRequest {
-	r.limit = &limit
-	return r
-}
-
-// Offset into the full result set. Usually specified with *limit* to paginate through the results. See [V3 API Standard Collection Parameters](https://developer.sailpoint.com/idn/api/standard-collection-parameters) for more information.
-func (r ApiListUserListRequest) Offset(offset int32) ApiListUserListRequest {
-	r.offset = &offset
-	return r
-}
-
-// If *true* it will populate the *X-Total-Count* response header with the number of results that would be returned if *limit* and *offset* were ignored.  Since requesting a total count can have a performance impact, it is recommended not to send **count&#x3D;true** if that value will not be used.  See [V3 API Standard Collection Parameters](https://developer.sailpoint.com/idn/api/standard-collection-parameters) for more information.
-func (r ApiListUserListRequest) Count(count bool) ApiListUserListRequest {
-	r.count = &count
-	return r
-}
-
-// Sort results using the standard syntax described in [V3 API Standard Collection Parameters](https://developer.sailpoint.com/idn/api/standard-collection-parameters#sorting-results) Sorting is supported for the following fields: **id, name, created, modified, type, attribute, value, source.id**
-func (r ApiListUserListRequest) Sorters(sorters []QuerySort) ApiListUserListRequest {
-	r.sorters = &sorters
-	return r
-}
-
-// Filter results using the standard syntax described in [V3 API Standard Collection Parameters](https://developer.sailpoint.com/idn/api/standard-collection-parameters#filtering-results) Filtering is supported for the following fields and operators:  **id**: *eq, in*  **name**: *eq, in, sw*  **type**: *eq, in*  **attribute**: *eq, in*  **value**: *eq, in, sw*  **source.id**: *eq, in*  **requestable**: *eq* **created**: *gt, lt, ge, le* **modified**: *gt, lt, ge, le*
-func (r ApiListUserListRequest) Filters(filters string) ApiListUserListRequest {
-	r.filters = &filters
-	return r
-}
-
-func (r ApiListUserListRequest) Execute() (*UserList, *http.Response, error) {
-	return r.ApiService.ListUsersExecute(r)
-}
-
-/*
-ListUsers Account Aggregation (File)
-
-Aggregates a delimited file for the given source.  This only works for file-based sources.
-
-	@param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
-	@param id
-	@return ApiListUserListRequest
-*/
-func (a *ListUserListApiService) ListUsers(ctx context.Context) ApiListUserListRequest {
-	return ApiListUserListRequest{
-		ApiService: a,
-		ctx:        ctx,
-	}
-}
-
-// Execute executes the request
-func (a *ListUserListApiService) ListUsersExecute(r ApiListUserListRequest) (*UserList, *http.Response, error) {
-	var (
-		localVarHTTPMethod  = http.MethodGet
-		localVarPostBody    interface{}
-		formFiles           []formFile
-		localVarReturnValue *UserList
-	)
-
-	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ApiListUserListRequest.TaskResults")
-	if err != nil {
-		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
-	}
-
-	localVarPath := localBasePath + "/cc/api/user/list"
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	if r.limit != nil {
-		parameterAddToQuery(localVarQueryParams, "limit", r.limit, "")
-	}
-	if r.offset != nil {
-		parameterAddToQuery(localVarQueryParams, "offset", r.offset, "")
-	}
-	if r.count != nil {
-		parameterAddToQuery(localVarQueryParams, "count", r.count, "")
-	}
-	sort, err := json.Marshal(r.sorters)
-	if err != nil {
-		return nil, nil, err
-	}
-	// q.Add("sort", string(sort))
-
-	if r.sorters != nil {
-		parameterAddToQuery(localVarQueryParams, "sort", string(sort), "")
-	}
-
-
-	if r.filters != nil {
-		parameterAddToQuery(localVarQueryParams, "filters", r.filters, "")
-	}
-
-	// to determine the Content-Type header
-	localVarHTTPContentTypes := []string{}
-
-	// set Content-Type header
-	localVarHTTPContentType := selectHeaderContentType(localVarHTTPContentTypes)
-	if localVarHTTPContentType != "" {
-		localVarHeaderParams["Content-Type"] = localVarHTTPContentType
-	}
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := selectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
-	if err != nil {
-		return localVarReturnValue, nil, err
-	}
-
-	localVarHTTPResponse, err := a.client.callAPI(req)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	localVarHTTPResponse.Body = ioutil.NopCloser(bytes.NewBuffer(localVarBody))
-	if err != nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := &GenericOpenAPIError{
-			body:  localVarBody,
-			error: localVarHTTPResponse.Status,
-		}
-		return localVarReturnValue, localVarHTTPResponse, newErr
-	}
-
-	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-	if err != nil {
-		newErr := &GenericOpenAPIError{
-			body:  localVarBody,
-			error: err.Error(),
-		}
-		return localVarReturnValue, localVarHTTPResponse, newErr
-	}
-
-	return localVarReturnValue, localVarHTTPResponse, nil
-}
